Move cert verifier decorator to its own module
